pool: remove unnecessary unboxing of file size
Motivation:
As we never use file size as object no reasons to use Long.

Modification:
remove unnecessary unboxing of file size

Result:
unnecessary objects are not created.

Acked-by: Paul Millar
Acked-by: Lea Morschel
Target: master
Require-book: no
Require-notes: no

+++ modules/dcache/src/main/java/org/dcache/pool/repository/meta/db/CacheRepositoryEntryImpl.java
@@ -103,7 +103,7 @@ public class CacheRepositoryEntryImpl implements ReplicaRecord
         } else {
 
             AccessTimeInfo accessTimeInfo = repository.getAccessTimeInfo().get(pnfsId.toString());
-            Long size = Long.valueOf(storageInfo.getLegacySize());
+            long size = storageInfo.getLegacySize();
             if (size == 0) {
                 try {
                     _size = _fileStore.getFileAttributeView(pnfsId).readAttributes().size();
@@ -115,7 +115,7 @@ public class CacheRepositoryEntryImpl implements ReplicaRecord
                 }
 
             } else {
-                _size = size.longValue();
+                _size = size;
             }
